test: share Matrix snapshot helpers
